added comment about possibly using LUA_PATH
instead of settings.basepath in the loadfile calls. And if you do that -- you might be able to use the 'require' method instead of 'loadfile' which, I think, does caching.

Heck, if the LUA_PATH env variable doesn't work out, perhaps adding settings.basepath to package.path and then using 'require' instead of 'loadfile' would be better for performance?

Not terribly certain if it would persist caching between requests to ATS though, so maybe it's a moot point.